+++ Calculadora TP/Aritmetica.h
@@ -52,7 +52,7 @@ float dividir(float numUno, float numDos);
  * \return float
  *
  */
-float factorialUno(long numUno);
+float factorialUno(float numUno);
 /** \brief Realiza la operacion del segundo numero
  *
  * \param numDos float Segundo numero al cual se le aplicara la funcion factorial
